Use absolute paths in overlays for charm location
Use absolute paths in the overlay to locate the charm under test.
This PR changes zaza to use a fully qualified directoy to point
at the charm under test rather than a relative one. This allows zaza
to work with Juju 2.7 *1

I think that switching away from relative paths has merits
irrespective of the Juju 2.7 issue. Given the overlay is generated
from a template at run time for each zaza deploy the contents are
really invisible to the user. The rendered overlay is stored in a
temporary directory created by a tempfile.TemporaryDirectory context
so the file is automatically removed and the user never sees it. If
the user were to see it they would see the contents are of the form:

```
applications:
  nova-cell-controller-cell2:
    charm: ../../../nova-cell-controller
```

which is confusing because the relative path is relative to the
bundle being overlayed not the location of the overlay. Switching
to an absolute path makes the rendered overlay clearer.

*1 https://bugs.launchpad.net/juju/+bug/1851817